move DeveloperPrimaryActionComponent to Developers namespace
