--- app/components/developer_primary_action_component.rb
@@ -1,10 +1,0 @@
-class DeveloperPrimaryActionComponent < ApplicationComponent
-  def initialize(user:, developer:)
-    @user = user
-    @developer = developer
-  end
-
-  def owner?
-    @user&.developer == @developer
-  end
-end
