+++ app/components/developers/primary_action_component.rb
@@ -1,0 +1,12 @@
+module Developers
+  class PrimaryActionComponent < ApplicationComponent
+    def initialize(user:, developer:)
+      @user = user
+      @developer = developer
+    end
+
+    def owner?
+      @user&.developer == @developer
+    end
+  end
+end
