+++ app/views/developers/show.html.erb
@@ -26,7 +26,7 @@
                   <% end %>
                 </div>
 
-                <%= render DeveloperPrimaryActionComponent.new(user: current_user, developer: @developer) %>
+                <%= render Developers::PrimaryActionComponent.new(user: current_user, developer: @developer) %>
               </div>
 
               <aside class="mt-8 lg:hidden">
